MDL-48242 completion criteria: timestamp does not get applied
The mark_complete() function is not defined as accepting a timestamp
parameter, but system wide, functions that call it, pass a time value
to be applied. This value should be accepted and honored, otherwise
when the value populated via cron tasks will be the cron time, rather
than the completion time.